refactor(iocuak): move isTaskKind to task module
