+++ packages/iocuak/src/task/utils/isTaskKind.ts
@@ -1,5 +1,5 @@
-import { TaskKind } from '../task/models/domain/TaskKind';
-import { TaskKindType } from '../task/models/domain/TaskKindType';
+import { TaskKind } from '../models/domain/TaskKind';
+import { TaskKindType } from '../models/domain/TaskKindType';
 
 export function isTaskKind(value: unknown): value is TaskKind {
   const taskKind: TaskKind = value as TaskKind;
